requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aio-pika==9.5.5
    # via -r requirements.in
aiodns==3.5.0
    # via ccxt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.12
    # via
    #   -r requirements.in
    #   ccxt
aiormq==6.8.1
    # via
    #   -r requirements.in
    #   aio-pika
aiosignal==1.3.2
    # via aiohttp
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
asyncio==3.4.3
    # via -r requirements.in
asyncpg==0.30.0
    # via -r requirements.in
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
bcrypt==4.3.0
    # via -r requirements.in
bidict==0.23.1
    # via python-socketio
black==25.1.0
    # via -r requirements.in
blinker==1.9.0
    # via streamlit
build==1.2.2.post1
    # via pip-tools
cachetools==5.5.2
    # via streamlit
ccxt==4.4.89
    # via -r requirements.in
certifi==2025.4.26
    # via
    #   ccxt
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   pycares
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   -r requirements.in
    #   black
    #   pip-tools
    #   streamlit
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
    #   pytest
cryptography==45.0.4
    # via
    #   -r requirements.in
    #   ccxt
dnspython==2.7.0
    # via pymongo
exceptiongroup==1.3.0
    # via aio-pika
fastapi==0.115.12
    # via -r requirements.in
flake8==7.2.0
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via opentelemetry-api
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.24.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
motor==3.7.1
    # via -r requirements.in
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
mypy==1.16.0
    # via -r requirements.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
narwhals==1.42.1
    # via altair
numpy==2.3.0
    # via
    #   -r requirements.in
    #   pandas
    #   pydeck
    #   streamlit
opentelemetry-api==1.34.1
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-jaeger-thrift
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asyncpg
    #   opentelemetry-instrumentation-pymongo
    #   opentelemetry-instrumentation-redis
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-jaeger-thrift==1.21.0
    # via -r requirements.in
opentelemetry-instrumentation==0.55b1
    # via
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asyncpg
    #   opentelemetry-instrumentation-pymongo
    #   opentelemetry-instrumentation-redis
opentelemetry-instrumentation-aiohttp-client==0.55b1
    # via -r requirements.in
opentelemetry-instrumentation-asyncpg==0.55b1
    # via -r requirements.in
opentelemetry-instrumentation-pymongo==0.55b1
    # via -r requirements.in
opentelemetry-instrumentation-redis==0.55b1
    # via -r requirements.in
opentelemetry-sdk==1.34.1
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-jaeger-thrift
opentelemetry-semantic-conventions==0.55b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asyncpg
    #   opentelemetry-instrumentation-pymongo
    #   opentelemetry-instrumentation-redis
    #   opentelemetry-sdk
opentelemetry-util-http==0.55b1
    # via opentelemetry-instrumentation-aiohttp-client
packaging==24.2
    # via
    #   altair
    #   black
    #   build
    #   opentelemetry-instrumentation
    #   pytest
    #   streamlit
pamqp==3.3.0
    # via aiormq
pandas==2.3.0
    # via
    #   -r requirements.in
    #   streamlit
pathspec==0.12.1
    # via
    #   black
    #   mypy
pika==1.3.2
    # via -r requirements.in
pillow==11.2.1
    # via streamlit
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.8
    # via black
pluggy==1.6.0
    # via pytest
prometheus-client==0.22.1
    # via -r requirements.in
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.1
    # via streamlit
psycopg2-binary==2.9.10
    # via -r requirements.in
pyarrow==20.0.0
    # via streamlit
pycares==4.9.0
    # via aiodns
pycodestyle==2.13.0
    # via flake8
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via -r requirements.in
pydeck==0.9.1
    # via streamlit
pyflakes==3.3.2
    # via flake8
pygments==2.19.1
    # via
    #   pytest
    #   rich
pymongo==4.13.1
    # via
    #   -r requirements.in
    #   motor
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.0
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-mock
pytest-asyncio==1.0.0
    # via -r requirements.in
pytest-mock==3.14.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   pandas
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   pydantic-settings
python-engineio==4.12.2
    # via python-socketio
python-socketio==5.13.0
    # via -r requirements.in
pytz==2025.2
    # via
    #   -r requirements.in
    #   pandas
redis==6.2.0
    # via -r requirements.in
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   ccxt
    #   streamlit
rich==14.0.0
    # via -r requirements.in
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
simple-websocket==1.1.0
    # via python-engineio
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
streamlit==1.45.1
    # via -r requirements.in
structlog==25.4.0
    # via -r requirements.in
tenacity==9.1.2
    # via streamlit
thrift==0.22.0
    # via opentelemetry-exporter-jaeger-thrift
toml==0.10.2
    # via streamlit
tornado==6.5.1
    # via streamlit
typing-extensions==4.14.0
    # via
    #   altair
    #   anyio
    #   ccxt
    #   exceptiongroup
    #   fastapi
    #   mypy
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   referencing
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
uvicorn==0.34.3
    # via -r requirements.in
watchdog==6.0.0
    # via streamlit
websockets==15.0.1
    # via -r requirements.in
wheel==0.45.1
    # via pip-tools
wrapt==1.17.2
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-redis
wsproto==1.2.0
    # via simple-websocket
yarl==1.20.1
    # via
    #   aio-pika
    #   aiohttp
    #   aiormq
    #   ccxt
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
